refactor(dialog-remote-s3): ConditionalSend ByteStream; cross-target blob test
Replace the per-target `ByteStream` alias (`+ Send` off the web, unbounded on
wasm) with a single `ByteChunks: Stream + ConditionalSend` supertrait alias, so
there is no `cfg` branching on the boxed stream type.

Move the blob import/read round-trip out of the native-only `#[tokio::test]`
unit test into the `tests/s3` integration suite (`#[dialog_common::test]`
against the real endpoint), so it runs on both native and wasm like the archive
and memory suites — adding a ranged read and a NotFound case.
