+++ rust/dialog-remote-s3/tests/s3/main.rs
@@ -4,5 +4,6 @@
 //! R2S3_ENDPOINT, R2S3_REGION, R2S3_BUCKET, R2S3_ACCESS_KEY_ID, R2S3_SECRET_ACCESS_KEY
 
 mod archive;
+mod blob;
 mod environment;
 mod memory;
